pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "reverso"
version = "0.1.0"
description = "Efficient time-series foundation models for zero-shot forecasting"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.6.0",
    "numpy",
    "pandas",
    "flash-linear-attention",
]

[project.optional-dependencies]
examples = [
    "matplotlib",
    "gluonts~=0.15.1",
    "python-dotenv>=1.0.0",
]

[project.urls]
Repository = "https://github.com/shinfxh/reverso"

[tool.setuptools.packages.find]
include = ["reverso*"]
